pyproject.toml
[tool.poetry]
name = "webtronics-mini-blog"
version = "0.1.0"
description = "test_case"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

fastapi = "0.91.0"
uvicorn = "^0.20.0"
sqlalchemy = "^2.0.3"
psycopg2-binary = "^2.9.5"
pydantic = {extras =["email"], version = "^1.10.4"}
alembic = "^1.9.3"

python-dotenv = "^0.21.1"

python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
types-passlib = "^1.7.7.8"
python-multipart = "^0.0.5"
bcrypt = "^4.0.1"

pytest = "^7.2.1"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]
black = "23.1.0"

isort = "^5.12.0"

mypy = "^1.0.0"

flake8 = "6.0.0"
flake8-black = "0.3.6"
flake8-builtins = "2.1.0"
flake8-isort = "6.0.0"

dotenv-linter = "^0.4.0"

pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
exclude = """
    ('
    |.idea/
    |.venv/
    |.vscode/
    ')
"""
